-generate PDF in lib utils
